altered structure of db manipulation
